requirements.txt
flask>=2.3.0
PyJWT>=2.8.0
Pillow>=10.0.0
werkzeug>=2.3.0
gunicorn>=21.2.0
